+++ src/hooks/api/db.ts
@@ -2,10 +2,9 @@ import IProduct, { ProductStatus } from "@/types/product";
 
 export const fakeDatabase: IProduct[] = Array.from({ length: 15 }).map(
   (_, index) => ({
-    id: index.toString(),
-    name: `Product ${index}`,
+    id: index.toString() + 1,
+    name: `Product ${index + 1}`,
     price: Math.random() * 100,
-    payLink: "https://www.google.com",
     status: Math.random() > 0.5 ? ProductStatus.Active : ProductStatus.Inactive,
   })
 );