+++ changelog.md
@@ -1,7 +1,13 @@
 ## Changelog
 
-### Update 1 - Monday 20/01/2024
+### Update 2 - Tuesday 23/01/2024
+#### Version v0.3.0:
 
+- Added update and delete products functionality+ react query logic to handle server updates
+- Updated create product drawer to handle update logic.
+- Created delete product modal + updated icon button design
+
+### Update 1 - Monday 22/01/2024
 #### Version v0.2.0:
 
 - Updated input components styles configuration to match design.